requirements.txt
python-telegram-bot[job-queue]==21.5
httpx==0.27.0
beautifulsoup4==4.12.3
lxml==5.2.2
